requirements.txt
numpy==1.23.0
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.12.1
--extra-index-url https://download.pytorch.org/whl/cu116
torchvision==0.13.1
Pillow==9.2.0
tqdm==4.64.0
matplotlib==3.5.3
torchmetrics==0.9.3
pandas==1.4.3
boto3==1.24.60
botocore==1.27.60
=======
wandb==0.13.2
timm==0.6.7
